fix(tests,ci): align ruff pre-commit pin with CI; satisfy newer lints
The pre-commit pin (ruff v0.6.9) and the dev-group ruff resolved by
``uv`` (0.15.11) had drifted apart. CI runs the newer ruff via
``uv run ruff check`` and flagged five issues not caught by
pre-commit on push:

- RUF043 in tests/test_manifest.py: regex match with unescaped ``.``.
  Switch to raw string.
- RUF046 in tests/evaluators/test_binary.py (×2): ``int(round(x[0]))``
  is redundant since ``round()`` already returns int. Drop the cast.
- RUF100 in tests/trainers/test_lightning_encoding.py (×2): ``# noqa: E402``
  is unused with ruff 0.11+ which recognises ``pytest.importorskip()``
  as a legitimate pre-import guard and stops flagging E402 for imports
  beneath it. Remove the noqa.

Also bump ``ruff-pre-commit`` to v0.15.11 so the pre-commit hook
matches what CI runs. Lint drift between pre-commit and CI was a
latent issue flagged in code reviews of the maldet 2.0 work — closing
it here so future PRs can't merge with new ruff lints.

The 2.0.0 PyPI release is unaffected (test/CI-only changes); main's
CI returns to green without bumping the version.

+++ tests/evaluators/test_binary.py
@@ -120,12 +120,12 @@ def test_metrics_correct_when_positive_class_is_index_zero() -> None:
         def extract(self, s: Sample) -> np.ndarray:
             # Each sample's feature is its label-as-index under
             # classes=["Malware","Benign"]: Malware=0, Benign=1. The model predicts
-            # ``int(round(x[0]))`` so trainer/evaluator round-trip is perfect.
+            # ``round(x[0])`` returns int; trainer/evaluator round-trip is perfect.
             return np.array([0.0 if s.label == "Malware" else 1.0], dtype=np.float32)
 
     class _PerfectModel:
         def predict(self, X: np.ndarray) -> np.ndarray:  # noqa: N803
-            return np.asarray([int(round(x[0])) for x in X], dtype=np.int64)
+            return np.asarray([round(x[0]) for x in X], dtype=np.int64)
 
     samples = [
         Sample(
@@ -172,7 +172,7 @@ def test_metrics_correct_when_positive_class_is_index_one() -> None:
 
     class _PerfectModel:
         def predict(self, X: np.ndarray) -> np.ndarray:  # noqa: N803
-            return np.asarray([int(round(x[0])) for x in X], dtype=np.int64)
+            return np.asarray([round(x[0]) for x in X], dtype=np.int64)
 
     samples = [
         Sample(
